Clear cache button

+++ src/db/DetailsDatabase.ts
@@ -10,23 +10,23 @@ interface DetailsRow {
 }
 
 export class DetailsDatabase extends Dexie {
-    private maxHours: number;
     public details!: Table<DetailsRow, number>;
 
-    public constructor(maxHours: number) {
+    public constructor(maxHours?: number) {
         super("DetailsDatabase");
-        this.maxHours = maxHours;
         this.version(1).stores({
             details: "++id, language"
         });
 
-        setTimeout(() => {
-            void this.transaction('rw', this.details, async () => {
-                const threshold = new Date(Date.now() - 1000 * 60 * 60 * this.maxHours),
-                    count = await this.details.filter(row => row.timestamp !== undefined && new Date(row.timestamp) < threshold).delete();
-                if (process.env.NODE_ENV === 'development') console.debug("Evicted old maps from indexedDB", { count, threshold });
-            });
-        }, 10_000);
+        if (maxHours) {
+            setTimeout(() => {
+                void this.transaction('rw', this.details, async () => {
+                    const threshold = new Date(Date.now() - 1000 * 60 * 60 * maxHours),
+                        count = await this.details.filter(row => row.timestamp !== undefined && new Date(row.timestamp) < threshold).delete();
+                    if (process.env.NODE_ENV === 'development') console.debug("Evicted old maps from indexedDB", { count, threshold });
+                });
+            }, 10_000);
+        }
     }
 
     public async getDetails(wikidataIDs: Set<string>, language?: string): Promise<Record<string, EtymologyDetails> | undefined> {
