+++ src/db/MapDatabase.ts
@@ -5,25 +5,25 @@ import type { BBox } from 'geojson';
 type MapRow = EtymologyResponse & { id?: number };
 
 export class MapDatabase extends Dexie {
-    private maxHours: number;
     public maps!: Table<MapRow, number>;
 
-    public constructor(maxHours: number) {
+    public constructor(maxHours?: number) {
         super("MapDatabase");
-        this.maxHours = maxHours;
 
         this.version(5).stores({
             //maps: "++id, [backEndID+onlyCentroids+language]" // Does not work: https://stackoverflow.com/a/56661425
             maps: "++id, [backEndID+language]"
         });
 
-        setTimeout(() => {
-            void this.transaction('rw', this.maps, async () => {
-                const threshold = new Date(Date.now() - 1000 * 60 * 60 * this.maxHours),
-                    count = await this.maps.filter(row => row.timestamp !== undefined && new Date(row.timestamp) < threshold).delete();
-                if (process.env.NODE_ENV === 'development') console.debug("Evicted old maps from indexedDB", { count, threshold });
-            });
-        }, 10_000);
+        if (maxHours) {
+            setTimeout(() => {
+                void this.transaction('rw', this.maps, async () => {
+                    const threshold = new Date(Date.now() - 1000 * 60 * 60 * maxHours),
+                        count = await this.maps.filter(row => row.timestamp !== undefined && new Date(row.timestamp) < threshold).delete();
+                    if (process.env.NODE_ENV === 'development') console.debug("Evicted old maps from indexedDB", { count, threshold });
+                });
+            }, 10_000);
+        }
     }
 
     public async getMap(backEndID: string, onlyCentroids: boolean, bbox: BBox, language?: string): Promise<MapRow | undefined> {
