+++ src/db/StatsDatabase.ts
@@ -12,23 +12,23 @@ interface StatsRow {
 }
 
 export class StatsDatabase extends Dexie {
-    private maxHours: number;
     public stats!: Table<StatsRow, number>;
 
-    public constructor(maxHours: number) {
+    public constructor(maxHours?: number) {
         super("StatsDatabase");
-        this.maxHours = maxHours;
         this.version(1).stores({
             stats: "++id, [colorSchemeID+language+wikidataIDs]"
         });
 
-        setTimeout(() => {
-            void this.transaction('rw', this.stats, async () => {
-                const threshold = new Date(Date.now() - 1000 * 60 * 60 * this.maxHours),
-                    count = await this.stats.filter(row => row.timestamp !== undefined && new Date(row.timestamp) < threshold).delete();
-                if (process.env.NODE_ENV === 'development') console.debug("Evicted old maps from indexedDB", { count, threshold });
-            });
-        }, 10_000);
+        if (maxHours) {
+            setTimeout(() => {
+                void this.transaction('rw', this.stats, async () => {
+                    const threshold = new Date(Date.now() - 1000 * 60 * 60 * maxHours),
+                        count = await this.stats.filter(row => row.timestamp !== undefined && new Date(row.timestamp) < threshold).delete();
+                    if (process.env.NODE_ENV === 'development') console.debug("Evicted old maps from indexedDB", { count, threshold });
+                });
+            }, 10_000);
+        }
     }
 
     public async getStats(colorSchemeID: ColorSchemeID, wikidataIDs: string[], language?: string): Promise<EtymologyStat[] | undefined> {
